+++ CMakeLists.txt
@@ -66,6 +66,7 @@ set(SOURCES
         engine/sorted_collection/skiplist.cpp
         engine/sorted_collection/rebuilder.cpp
         engine/hash_collection/hash_list.cpp
+        engine/list_collection/list.cpp
         engine/write_batch_impl.cpp
         engine/dram_allocator.cpp
         engine/pmem_allocator/pmem_allocator.cpp
