+++ engine/list_collection/list.hpp
@@ -24,6 +24,11 @@
 namespace KVDK_NAMESPACE {
 class ListIteratorImpl;
 
+enum class ListPos : int {
+  Front = 0,
+  Back = -1,
+};
+
 class List : public Collection {
  public:
   List(DLRecord* header, const StringView& name, CollectionIDType id,
@@ -56,7 +61,7 @@ class List : public Collection {
     Status s{Status::InvalidArgument};
     std::vector<SpaceEntry> spaces;
     std::vector<StringView> elems;
-    int pos{0};
+    ListPos pos;
     TimeStampType ts;
   };
 
@@ -73,376 +78,41 @@ class List : public Collection {
   bool HasExpired() const final { return dl_list_.Header()->HasExpired(); }
 
   WriteResult SetExpireTime(ExpireTimeType expired_time,
-                            TimeStampType timestamp) {
-    WriteResult ret;
-    DLRecord* header = HeaderRecord();
-    SpaceEntry space = pmem_allocator_->Allocate(
-        DLRecord::RecordSize(header->Key(), header->Value()));
-    if (space.size == 0) {
-      ret.s = Status::PmemOverflow;
-      return ret;
-    }
-    DLRecord* pmem_record = DLRecord::PersistDLRecord(
-        pmem_allocator_->offset2addr_checked(space.offset), space.size,
-        timestamp, RecordType::ListRecord, RecordStatus::Normal,
-        pmem_allocator_->addr2offset_checked(header), header->prev,
-        header->next, header->Key(), header->Value(), expired_time);
-    bool success = dl_list_.Replace(header, pmem_record);
-    kvdk_assert(success, "existing header should be linked on its list");
-    ret.existing_record = header;
-    ret.write_record = pmem_record;
-    return ret;
-  }
+                            TimeStampType timestamp);
 
-  Status BatchPopBack(size_t n, std::vector<std::string>* elems,
-                      TimeStampType batch_ts, char* batch_log) {
-    return batchPopImpl(n, -1, elems, batch_ts, batch_log);
-  }
+  WriteResult PushFront(const StringView& elem, TimeStampType ts);
 
-  Status BatchPopFront(size_t n, std::vector<std::string>* elems,
-                       TimeStampType batch_ts, char* batch_log) {
-    return batchPopImpl(n, 0, elems, batch_ts, batch_log);
-  }
+  WriteResult PushBack(const StringView& elem, TimeStampType ts);
 
-  WriteResult PushFront(const StringView& elem, TimeStampType ts,
-                        const SpaceEntry space) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
+  WriteResult PopFront(TimeStampType ts);
 
-    DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                           RecordStatus::Normal, ts, space);
-    ret.s = dl_list_.PushFront(args);
-    kvdk_assert(ret.s == Status::Ok, "Push front should alwasy success");
-    UpdateSize(1);
-    ret.write_record =
-        pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-    return ret;
-  }
-
-  WriteResult PushFront(const StringView& elem, TimeStampType ts) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
-    SpaceEntry space =
-        pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-    if (space.size == 0) {
-      ret.s = Status::PmemOverflow;
-      return ret;
-    }
-
-    DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                           RecordStatus::Normal, ts, space);
-    ret.s = dl_list_.PushFront(args);
-    kvdk_assert(ret.s == Status::Ok, "Push front should alwasy success");
-    UpdateSize(1);
-    ret.write_record =
-        pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-    return ret;
-  }
-
-  WriteResult PushBack(const StringView& elem, TimeStampType ts) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
-    SpaceEntry space =
-        pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-    if (space.size == 0) {
-      ret.s = Status::PmemOverflow;
-      return ret;
-    }
-
-    DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                           RecordStatus::Normal, ts, space);
-    ret.s = dl_list_.PushBack(args);
-    kvdk_assert(ret.s == Status::Ok, "Push front should alwasy success");
-    UpdateSize(1);
-    ret.write_record =
-        pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-    return ret;
-  }
-
-  WriteResult PushBack(const StringView& elem, TimeStampType ts,
-                       const SpaceEntry& space) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
-
-    DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                           RecordStatus::Normal, ts, space);
-    ret.s = dl_list_.PushBack(args);
-    kvdk_assert(ret.s == Status::Ok, "Push back should alwasy success");
-    UpdateSize(1);
-    ret.write_record =
-        pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-    return ret;
-  }
-
-  WriteResult PopFront(TimeStampType ts) {
-    WriteResult ret;
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (iter.SeekToFirst(); iter.Valid(); iter.Next()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal) {
-        SpaceEntry space =
-            pmem_allocator_->Allocate(DLRecord::RecordSize(record->Key(), ""));
-        if (space.size == 0) {
-          ret.s = Status::PmemOverflow;
-          return ret;
-        }
-        DLList::WriteArgs args(record->Key(), "", RecordType::ListElem,
-                               RecordStatus::Outdated, ts, space);
-        ret.s = dl_list_.Update(args, record);
-        kvdk_assert(ret.s == Status::Ok,
-                    "the whole list is locked so the update must be success");
-        ret.write_record =
-            pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-        ret.existing_record = record;
-        UpdateSize(-1);
-        return ret;
-      }
-    }
-    ret.s = Status::NotFound;
-    return ret;
-  };
-
-  WriteResult PopBack(TimeStampType ts) {
-    // TODO cache back to avoid iter
-    WriteResult ret;
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (iter.SeekToLast(); iter.Valid(); iter.Prev()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal) {
-        SpaceEntry space =
-            pmem_allocator_->Allocate(DLRecord::RecordSize(record->Key(), ""));
-        if (space.size == 0) {
-          ret.s = Status::PmemOverflow;
-          return ret;
-        }
-        DLList::WriteArgs args(record->Key(), "", RecordType::ListElem,
-                               RecordStatus::Outdated, ts, space);
-        ret.s = dl_list_.Update(args, record);
-        kvdk_assert(ret.s == Status::Ok,
-                    "the whole list is locked so the update must be success");
-        ret.write_record =
-            pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-        ret.existing_record = record;
-        UpdateSize(-1);
-        return ret;
-      }
-    }
-    ret.s = Status::NotFound;
-    return ret;
-  }
+  WriteResult PopBack(TimeStampType ts);
 
   WriteResult InsertBefore(const StringView& elem,
-                           const StringView& existing_elem, TimeStampType ts) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (iter.SeekToFirst(); iter.Valid(); iter.Next()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal &&
-          equal_string_view(record->Value(), existing_elem)) {
-        SpaceEntry space =
-            pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-        if (space.size == 0) {
-          ret.s = Status::PmemOverflow;
-          return ret;
-        }
-        DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                               RecordStatus::Normal, ts, space);
-        ret.s = dl_list_.InsertBefore(args, record);
-        kvdk_assert(
-            ret.s == Status::Ok,
-            "the whole list is locked, so the insertion must be success");
-        if (ret.s == Status::Ok) {
-          ret.write_record =
-              pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-          UpdateSize(1);
-        }
-        return ret;
-      }
-    }
-    ret.s = Status::NotFound;
-    return ret;
-  }
+                           const StringView& existing_elem, TimeStampType ts);
 
   WriteResult InsertAfter(const StringView& elem,
-                          const StringView& existing_elem, TimeStampType ts) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (iter.SeekToFirst(); iter.Valid(); iter.Next()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal &&
-          equal_string_view(record->Value(), existing_elem)) {
-        SpaceEntry space =
-            pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-        if (space.size == 0) {
-          ret.s = Status::PmemOverflow;
-          return ret;
-        }
-        DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                               RecordStatus::Normal, ts, space);
-        ret.s = dl_list_.InsertAfter(args, record);
-        kvdk_assert(
-            ret.s == Status::Ok,
-            "the whole list is locked, so the insertion must be success");
-        if (ret.s == Status::Ok) {
-          ret.write_record =
-              pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-          UpdateSize(1);
-        }
-        return ret;
-      }
-    }
-    ret.s = Status::NotFound;
-    return ret;
-  }
+                          const StringView& existing_elem, TimeStampType ts);
 
-  WriteResult InsertAt(const StringView& elem, uint64_t pos, TimeStampType ts) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    uint64_t cur = 0;
-    DLRecord* prev = dl_list_.Header();
-    for (iter.SeekToFirst(); iter.Valid() && cur < pos; iter.Next()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Outdated) {
-        continue;
-      }
-      cur++;
-      prev = record;
-    }
+  WriteResult InsertAt(const StringView& elem, uint64_t pos, TimeStampType ts);
 
-    if (cur < pos) {
-      ret.s = Status::NotFound;
-      return ret;
-    }
+  WriteResult Erase(uint64_t pos);
 
-    SpaceEntry space =
-        pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-    if (space.size == 0) {
-      ret.s = Status::PmemOverflow;
-      return ret;
-    }
-    DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                           RecordStatus::Normal, ts, space);
-    ret.s = dl_list_.InsertAfter(args, prev);
-    kvdk_assert(ret.s == Status::Ok,
-                "the whole list is locked, so the insertion must be success");
-    ret.write_record =
-        pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-    UpdateSize(1);
-    return ret;
-  }
+  Status Front(std::string* elem);
 
-  WriteResult Erase(uint64_t pos) {
-    WriteResult ret;
-    if (pos >= Size()) {
-      ret.s = Status::NotFound;
-      return ret;
-    }
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    uint64_t cur = 0;
-    for (iter.SeekToFirst(); iter.Valid(); iter.Next()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Outdated) {
-        continue;
-      }
-      if (cur == pos) {
-        bool success = dl_list_.Remove(record);
-        ret.existing_record = record;
-        kvdk_assert(success,
-                    "the whole list is locked, so the remove must be success");
-        break;
-      }
-      cur++;
-    }
-    kvdk_assert(cur == pos, "size already checked");
-    UpdateSize(-1);
-    return ret;
-  }
-
-  Status Front(std::string* elem) {
-    if (Size() == 0) {
-      return Status::NotFound;
-    }
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (iter.SeekToFirst(); iter.Valid(); iter.Next()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal) {
-        StringView sw = record->Value();
-        elem->assign(sw.data(), sw.size());
-        return Status::Ok;
-      }
-    }
-
-    return Status::NotFound;
-  }
-
-  Status Back(std::string* elem) {
-    if (Size() == 0) {
-      return Status::NotFound;
-    }
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (iter.SeekToLast(); iter.Valid(); iter.Prev()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal) {
-        StringView sw = record->Value();
-        elem->assign(sw.data(), sw.size());
-        return Status::Ok;
-      }
-    }
-
-    return Status::NotFound;
-  }
+  Status Back(std::string* elem);
 
   bool Replace(DLRecord* old_record, DLRecord* new_record) {
     return dl_list_.Replace(old_record, new_record);
   }
 
-  WriteResult Replace(uint64_t pos, const StringView& elem, TimeStampType ts) {
-    WriteResult ret;
-    std::string internal_key(InternalKey(""));
-    if (pos >= Size()) {
-      ret.s = Status::NotFound;
-      return ret;
-    }
-    SpaceEntry space =
-        pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-    if (space.size == 0) {
-      ret.s = Status::PmemOverflow;
-      return ret;
-    }
-    DLList::WriteArgs args(internal_key, elem, RecordType::ListElem,
-                           RecordStatus::Normal, ts, space);
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    uint64_t cur = 0;
-    for (iter.SeekToFirst(); iter.Valid(); iter.Next()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Outdated) {
-        continue;
-      }
-      if (cur == pos) {
-        ret.s = dl_list_.Update(args, record);
-        ret.existing_record = record;
-        ret.write_record =
-            pmem_allocator_->offset2addr_checked<DLRecord>(space.offset);
-        kvdk_assert(ret.s == Status::Ok,
-                    "the whole list is locked, so the update must be success");
-        break;
-      }
-      cur++;
-    }
-    kvdk_assert(cur == pos, "size already checked");
-    return ret;
-  }
+  WriteResult Update(uint64_t pos, const StringView& elem, TimeStampType ts);
 
   void UpdateSize(int64_t delta) {
     kvdk_assert(delta >= 0 || size_.load() >= static_cast<size_t>(-delta),
                 "Update hash list size to negative");
     size_.fetch_add(delta, std::memory_order_relaxed);
   }
-
-  Status Move() { return Status::Ok; }
 
   size_t Size() { return size_.load(); }
 
@@ -456,100 +126,15 @@ class List : public Collection {
 
   void Destroy() {}
 
-  PushNArgs PreparePushN(int pos, const std::vector<StringView>& elems,
-                         TimeStampType ts) {
-    PushNArgs args;
-    args.pos = pos;
-    args.ts = ts;
-    if (elems.size() > 0) {
-      std::string internal_key(InternalKey(""));
-      for (auto& elem : elems) {
-        SpaceEntry space =
-            pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-        if (space.size == 0) {
-          GlobalLogger.Error("Try allocate %lu error\n",
-                             DLRecord::RecordSize(internal_key, elem));
-          for (auto& sp : args.spaces) {
-            pmem_allocator_->Free(sp);
-          }
-          args.s = Status::PmemOverflow;
-          break;
-        }
-        args.spaces.emplace_back(space);
-      }
-      args.elems = elems;
-    }
-    args.s = Status::Ok;
-    return args;
-  }
+  PushNArgs PreparePushN(ListPos pos, const std::vector<StringView>& elems,
+                         TimeStampType ts);
 
-  PopNArgs PreparePopN(size_t n, int pos, TimeStampType ts,
-                       std::vector<std::string>* elems) {
-    size_t nn = n;
-    PopNArgs args;
-    args.ts = ts;
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (pos == 0 ? iter.SeekToFirst() : iter.SeekToLast();
-         iter.Valid() && nn > 0; pos == 0 ? iter.Next() : iter.Prev()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal) {
-        SpaceEntry space =
-            pmem_allocator_->Allocate(DLRecord::RecordSize(record->Key(), ""));
-        if (space.size == 0) {
-          for (auto& sp : args.spaces) {
-            pmem_allocator_->Free(sp);
-          }
-          args.s = Status::PmemOverflow;
-          return args;
-        }
-        if (elems) {
-          StringView sw = record->Value();
-          elems->emplace_back(sw.data(), sw.size());
-        }
-        args.spaces.emplace_back(space);
-        args.to_pop.emplace_back(record);
-        nn--;
-      }
-    }
-    args.s = Status::Ok;
-    return args;
-  }
+  PopNArgs PreparePopN(ListPos pos, size_t n, TimeStampType ts,
+                       std::vector<std::string>* elems);
 
-  Status PushN(const PushNArgs& args) {
-    if (args.s != Status::Ok) {
-      return args.s;
-    }
-    std::string internal_key(InternalKey(""));
-    kvdk_assert(args.elems.size() == args.spaces.size(), "");
-    for (size_t i = 0; i < args.elems.size(); i++) {
-      DLList::WriteArgs wa(internal_key, args.elems[i], RecordType::ListElem,
-                           RecordStatus::Normal, args.ts, args.spaces[i]);
-      Status s = args.pos == 0 ? dl_list_.PushFront(wa) : dl_list_.PushBack(wa);
-      kvdk_assert(s == Status::Ok, "Push back/front should always success");
-      TEST_CRASH_POINT("List::PushN", "");
-    }
-    UpdateSize(args.elems.size());
-    return Status::Ok;
-  }
+  Status PushN(const PushNArgs& args);
 
-  Status PopN(const PopNArgs& args) {
-    if (args.s != Status::Ok) {
-      return args.s;
-    }
-    std::string internal_key(InternalKey(""));
-    kvdk_assert(args.spaces.size() == args.to_pop.size(), "");
-    for (size_t i = 0; i < args.to_pop.size(); i++) {
-      DLList::WriteArgs wa(internal_key, "", RecordType::ListElem,
-                           RecordStatus::Outdated, args.ts, args.spaces[i]);
-      Status s = dl_list_.Update(wa, args.to_pop[i]);
-      kvdk_assert(
-          s == Status::Ok,
-          "the whole list should be locked, so the update must be success");
-      TEST_CRASH_POINT("List::PopN", "");
-    }
-    UpdateSize(-args.to_pop.size());
-    return Status::Ok;
-  }
+  Status PopN(const PopNArgs& args);
 
   static CollectionIDType ListID(DLRecord* record) {
     assert(record != nullptr);
@@ -567,77 +152,6 @@ class List : public Collection {
   }
 
  private:
-  Status batchPushImpl(int pos, const std::vector<StringView>& elems,
-                       TimeStampType batch_ts, char* batch_log) {
-    BatchWriteLog log;
-    log.SetTimestamp(batch_ts);
-    std::vector<DLList::WriteArgs> write_args;
-    std::string internal_key(InternalKey(""));
-    for (auto& elem : elems) {
-      SpaceEntry space =
-          pmem_allocator_->Allocate(DLRecord::RecordSize(internal_key, elem));
-      if (space.size == 0) {
-        return Status::PmemOverflow;
-      }
-      write_args.emplace_back(internal_key, elem, RecordType::ListElem,
-                              RecordStatus::Normal, batch_ts, space);
-      log.ListEmplace(space.offset);
-    }
-    log.EncodeTo(batch_log);
-    BatchWriteLog::MarkProcessing(batch_log);
-    for (auto& wa : write_args) {
-      Status s = pos == 0 ? dl_list_.PushFront(wa) : dl_list_.PushBack(wa);
-      kvdk_assert(s == Status::Ok, "Push back should always success");
-    }
-    UpdateSize(elems.size());
-    BatchWriteLog::MarkCommitted(batch_log);
-    return Status::Ok;
-  }
-
-  Status batchPopImpl(size_t n, int pos, std::vector<std::string>* elems,
-                      TimeStampType batch_ts, char* batch_log) {
-    kvdk_assert(batch_log != nullptr, "");
-    BatchWriteLog log;
-    log.SetTimestamp(batch_ts);
-    if (elems) {
-      elems->clear();
-    }
-    size_t nn = n;
-    std::vector<DLRecord*> to_pop;
-    std::vector<DLList::WriteArgs> write_args;
-    DLListRecordIterator iter(&dl_list_, pmem_allocator_);
-    for (pos == 0 ? iter.SeekToFirst() : iter.SeekToLast();
-         iter.Valid() && nn-- > 0; pos == 0 ? iter.Next() : iter.Prev()) {
-      DLRecord* record = iter.Record();
-      if (record->GetRecordStatus() == RecordStatus::Normal) {
-        SpaceEntry space =
-            pmem_allocator_->Allocate(DLRecord::RecordSize(record->Key(), ""));
-        if (space.size == 0) {
-          return Status::PmemOverflow;
-        }
-        to_pop.emplace_back(record);
-        write_args.emplace_back(record->Key(), "", RecordType::ListElem,
-                                RecordStatus::Outdated, batch_ts, space);
-        log.ListDelete(space.offset);
-      }
-    }
-    log.EncodeTo(batch_log);
-    BatchWriteLog::MarkProcessing(batch_log);
-    for (size_t i = 0; i < to_pop.size(); i++) {
-      if (elems) {
-        StringView sw = to_pop[i]->Value();
-        elems->emplace_back(sw.data(), sw.size());
-      }
-      Status s = dl_list_.Update(write_args[i], to_pop[i]);
-      kvdk_assert(
-          s == Status::Ok,
-          "the whole list should be locked, so the update must be success");
-    }
-    UpdateSize(-to_pop.size());
-    BatchWriteLog::MarkCommitted(batch_log);
-    return Status::Ok;
-  }
-
   friend ListIteratorImpl;
   std::recursive_mutex list_lock_;
   DLList dl_list_;
@@ -653,8 +167,7 @@ class ListIteratorImpl final : public ListIterator {
         list_(list),
         snapshot_(snapshot),
         own_snapshot_(own_snapshot),
-        dl_iter_(&list->dl_list_, list->pmem_allocator_, snapshot,
-                 own_snapshot) {}
+        dl_iter_(&list->dl_list_, list->pmem_allocator_, snapshot) {}
 
   void Seek(long index) final {
     if (index < 0) {
@@ -730,6 +243,6 @@ class ListIteratorImpl final : public ListIterator {
   List* list_;
   const SnapshotImpl* snapshot_;
   bool own_snapshot_;
-  DLListAccessIterator dl_iter_;
+  DLListDataIterator dl_iter_;
 };
 }  // namespace KVDK_NAMESPACE
